more speed : everything to short

+++ finale/Localisation.cs
@@ -4,10 +4,10 @@ namespace finale
 {
 	public struct Localisation
 	{
-		public int Line { get; private set;}
-		public int Col { get; private set;}
+		public short Line { get; private set;}
+		public short Col { get; private set;}
 
-		public Localisation (int line, int col):this()
+		public Localisation (short line, short col):this()
 		{
 			Line = line;
 			Col = col;
